    \name{sim.data}
     \docType{data}
     \alias{sim.data}
     \title{Simulated data set of 500 observations}
     \description{This is a small dataset of 500 observations according to the \eqn{\beta_{jk}} parameterization. The number of reference variables is 1 and correspond to the values in the first column of the array \code{sim.data}. The number of response variables is 6 and correspond to the values in the last 6 column of the array \code{sim.data}. There are \eqn{J=3} conditions, while the number of replicates per condition is \eqn{L=(3,2,1)}.}
     \usage{sim.data}
     \format{A numeric array containing \eqn{500\times 7} observations.}
     \keyword{datasets}